Refactoring: re-organizing library layout
